+++ client/src/types/SceneComponent.ts
@@ -3,7 +3,6 @@ import type { Scene } from "three";
 export interface SceneComponent {
   add(scene: Scene): void;
   remove(scene: Scene): void;
-  animate(): void;
-  freeze(): void;
+  tick(deltaTime: number): void;
   setPosition(x: number, y: number, z: number): void;
 }